requirements.txt
requests
python-dotenv

# GIVE RECOMMENDATION__________________________
#print("-----------------------------------")
#if (  float(latest_close) > float((.2 * close_av) + close_av)  ):
    #print("RECOMMENDATION: Buy!")
    #print("REASONING: The latest closing price is 20% higher than the average closing price.")
#else:
    #print("RECOMMENDATION: DON'T BUY!")
    #print("REASON: Because the latest closing price is not 20% higher than the average closing price.")
#print("-----------------------------------")
#print("Writing info to csv" + str(csv_file_path))